requirements.txt
streamlit
requests
neo4j
pydantic
python-dotenv
pytest
